refactor: move adapter with new namespacing
